Cloudinary, db and carousel
Images are now hosted on cloudinary once a user creates a new post. The db is also updated with the models required for that.

Posts are now displayed in the users profile and in their feed. Furthermore, posts with several images are contained in a carousel so users can swipe through the collection