Conversation models
Scrapped and rebuilt the entire prisma schema for the dm functions of this application. Instead of separating conversations into two separate tables I merged them into one called conversations and also messages. So by this design I could remove several tables that like more or less identical for example GroupMessages and PrivateMessages. Furhtermore I could also delete the groupchat table since that logic is handled inside the conversationmember table

So all in all, mostly backend updates that are not seen in the code.