Realtime message
Needed to rebuild the entire backend that handles the messages in conversations since I changed the prisma model yesterday. So mostly removed and cleaned up the routes in the conversationRoute.js to fit the new tables. As of right now messages are inserted in both private and group conversations, but not updated in realtime using Supabase which is the next step