Adding and removing conversation members
Removed the old route (groupRoute.js) and migrated those functions associated with that controller (groupControllers.js) to conversationRoute.js and conversationController.js, since I recently updated the whole prisma schema and merged the conversations into one table instead of one for private and one for group. So now users can add users after a group conversation is created and kick them as long as they are the admin (creator) of that group. A smaller tweak is also so that the group name is updated based on the names in the group

Next step is to get supabase to provide me with the realtime updates, right now I cant seem to get the policy to let me through and read the data in a safe manner