Realtime message
Messages are now sent and received in realtime. The issue was that I had not granted the authenticated users access on the schema to query the tables, which is why it worked without RLS and not with

So now users can chat in realtime. Furthermore are also conversations removed in DirectMessages.jsx if the users is kicked from them or the conversation is deleted. Conversations are also updated if a user is invited to a conversation so they see it in realtime.

So now the project is more or less done. Final step is documentation and styling it so it works for phones